refactor(search): config for app
