+++ snippets/src/scripts/generate-id-mapper.ts
@@ -2,7 +2,7 @@ import fs from "fs-extra";
 import path from "path";
 
 import { MarkdownFileProcessor, processMarkdownDirectories, toIdMapperCollection } from "../libs/markdown-files";
-import { config } from "./_config";
+import { config } from "../_config";
 import { ConsoleLogger, Logger } from "../utils/logger";
 import { pinoLogBuilder } from "../utils/pino-log";
